Update FamiPong.ino
Refactored code to transition to using the AccelStepper library. This provides the ability to run the motors at higher speeds, and run them asynchronously without blocking other code running.

Also increased the I2C rate to the shield up to fast mode 400Khz to improve communication with motors through the stepper shield, to increase motor speeds further.

Added test code to run the motors back and forth with the Refactor, which is working well.